requirements-test.txt
mock
pytest
pytest-cov
pytest-django
pytest-sugar
